// src/app/models/user.ts
// export interface User {
//     id: string;
//     username: string;
//     password: string;
//     email: string;
//     firstName: string;
//     lastName: string;
//     token: string;
// }

export interface SignUpInfo {
    email: string;
    password: string;
    firstName: string;
    lastName: string;
}

export interface LoginInfo {
    email: string;
    password: string;
}